backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
azure-identity==1.15.0
python-dotenv==1.0.0
pydantic==2.10.2
redis
google-cloud-pubsub==2.33.0